Extracted to its on file logger  configs
